+++ src/main/java/com/baechu/book/service/BookService.java
@@ -1,7 +1,9 @@
 package com.baechu.book.service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -16,6 +18,25 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class BookService {
 	private final BookRepository bookRepository;
+
+
+	public Map<String, Object> bookdetail(Long bookid){
+		Map<String, Object> info = new HashMap<>();
+		Book book = bookRepository.findById(bookid).orElseThrow(
+			()-> new IllegalArgumentException("해당 번호의 책 없음")
+		);
+		info.put("bookid",bookid);
+		info.put("title", book.getTitle());
+		info.put("image", book.getImage());
+		info.put("price", book.getPrice());
+		info.put("author", book.getAuthor());
+		info.put("publish", book.getPublish());
+		String birth = book.getYear()+"년 "+book.getMonth()+"월";
+		info.put("birth", birth);
+		info.put("inventory", 7);
+
+		return info;
+	}
 
 	public List<BookDto> searchByWord(String query) {
 		List<BookDto> dtos = new ArrayList<>();
